pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "azwrap"
version = "0.1.7"
description = "Azure deployment wrapper"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "azure-ai-projects>=1.0.0b7",
    "azure-ai-resources>=1.0.0b9",
    "azure-ai-formrecognizer>=3.3.0",  # Required for Document Intelligence client
    "azure-identity>=1.21.0",
    "azure-mgmt-cognitiveservices>=13.6.0",
    "azure-mgmt-resource>=22.0.0",
    "azure-mgmt-search>=9.1.0",
    "azure-mgmt-storage>=22.1.1",
    "azure-search-documents>=11.5.2",
    "azure-storage-blob>=12.13.0",  # Required for blob storage operations
    "azure-data-tables==12.7.0",  # Required for table storage operations
    "azure-core>=1.24.0",  # Required for credentials
    "langchain>=0.3.20",
    "langchain-community>=0.3.19",
    "langchain-openai>=0.3.8",
    "langchain-core>=0.1.0",  # Required for core langchain functionality
    "langchain_text_splitters>=0.0.1",  # Required for text splitting
    "pillow>=11.2.1",
    "PyMuPDF>=1.26.0",
    "python-dotenv>=1.0.1",
    "PyPDF2>=3.0.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",  # Required for pandas Excel support
    "click>=8.1.8",     # Required for CLI
    "python-docx>=1.1.2",  # Required for DOCX file processing
    "regex>=2022.1.18",  # Required for regex operations
    "tenacity>=8.0.1",  # Required for retry functionality
    "tiktoken>=0.9.0",  # Required for chunk token count
    "openai>=1.0.0",    # Required for Azure OpenAI client
]

[project.scripts]
azwrap = "azwrap.main:main"

[tool.setuptools]
packages = ["azwrap"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",  # For running tests in test_identity.py
]
